src/main/java/mx/com/adurand/pruebabackend/controller/PriceController.java
package mx.com.adurand.pruebabackend.controller;

import mx.com.adurand.pruebabackend.model.dto.PriceDto;
import mx.com.adurand.pruebabackend.service.PriceService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.time.LocalDateTime;

@RestController
@RequestMapping("/price")
public class PriceController {

    private final PriceService priceService;

    public PriceController(PriceService priceService){
        this.priceService = priceService;
    }

    @GetMapping
    public PriceDto getPrice(@RequestParam int brandId, @RequestParam int productId,
                             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime date) {
        return priceService.getPrice(brandId, productId, date);
    }
}
